requirements.txt
discord.py>=1.6
geopy
ChatterBot
chatterbot-corpus
emoji
PyYAML
git+git://github.com/niklasf/python-chess.git
git+https://github.com/Rapptz/discord-ext-menus
CairoSVG
aiosqlite